fix: error of successfull registration
