+++ src/pages/main/index.ts
@@ -1,1 +1,1 @@
-export { Main } from "./main";
+export { Main } from "./ui";
