+++ src/pages/main/ui/style.css
@@ -1,0 +1,13 @@
+.main {
+  width: 100%;
+  height: 100vh;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  row-gap: 10px;
+  color: rgb(42, 14, 87);
+  background-image: url("https://images.unsplash.com/photo-1520340356584-f9917d1eea6f?fm=jpg&q=60&w=3000&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTB8fGNhciUyMGJhY2tncm91bmR8ZW58MHx8MHx8fDA%3D");
+  background-size: auto;
+}
