+++ src/pages/registration/ui/form/index.ts
@@ -1,6 +1,2 @@
 export { RegistrationForm } from "./form";
-export {
-  RegistrationSuccessMessage,
-  goAnimationAlert,
-  removeAnimationAlert,
-} from "./registration-success-message";
+export { RegistrationSuccessMessage } from "./registration-success-message";
